embedding-service/requirements.txt
fastapi==0.117.1
huggingface-hub==0.35.1
numpy==2.3.3
pillow==11.3.0
pydantic==2.11.9
python-multipart==0.0.20
requests==2.32.4
torch==2.8.0
torchvision==0.23.0
transformers==4.56.2
uvicorn==0.37.0
